+++ src/Result.php
@@ -82,7 +82,7 @@ final class Result
     {
         $errors = [];
         foreach ($this->errors as $error) {
-            $stringValuePath = implode($separator, $error->getValuePath(true));
+            $stringValuePath = implode($separator, $error->getValuePath($separator));
             $errors[$stringValuePath][] = $error->getMessage();
         }
 
@@ -172,7 +172,7 @@ final class Result
                 continue;
             }
 
-            $valuePath = implode($separator, array_slice($error->getValuePath(true), 1));
+            $valuePath = implode($separator, array_slice($error->getValuePath($separator), 1));
             $errors[$valuePath][] = $error->getMessage();
         }